Create GasLimit.sol
This Gas contract contains a 'function' named 'Forever' that attempts to run an infinite loop. The purpose of this loop is to consume all the gas supplied to the transaction.

1. 'uint public i = 0;': This declares a state variable 'i' of type 'uint' (unsigned integer) and initializes it with the value '0'.

2. 'function Forever() public {...}': This is the Forever function, marked as public, meaning it can be called externally. The purpose of this function is to run an infinite loop '(while (true))' that increments the 'i' variable.
   - The 'i += 1;' statement inside the loop increments the value of i indefinitely.

This contract, specifically the Forever function, is a demonstration of how running out of gas during the execution of a transaction can lead to the failure of that transaction, and any state changes made up to that point are reverted. It's important to use gas responsibly and avoid unintentional infinite loops in smart contracts to prevent transactions from getting stuck and consuming unnecessary resources.