Create ConditionalStatementForWhile.sol
The for loop initializes a variable 'i' to 0 and iterates as long as 'i' is less than 10. It includes conditional statements ('if') inside the loop, where if 'i' is equal to 3, it skips to the next iteration using 'continue', and if 'i' is equal to 5, it exits the loop using 'break'.

The while loop initializes a variable 'j' and increments it as long as 'j' is less than 10.

The 'pure' modifier indicates that the function will not modify the state of the contract.