Create Events.sol
This contract demonstrates the declaration and emission of events in Solidity. Events are a way for contracts to communicate information to the external world or to trigger specific actions in response to certain conditions.

Event Declaration:
   - Declares two events: 'Log' and 'AnotherLog'.
   - The 'Log' event has two parameters, 'address sender' and 'string' message. The 'address sender' parameter is marked as 'indexed', which means that it can be used for log filtering.
   - The 'AnotherLog' event has no parameters.

Event Emission:
   - The 'test' function is a public function that emits instances of the declared events.
   - It emits two instances of the 'Log' event with different messages and the same sender ('msg.sender').
   - It also emits an instance of the 'AnotherLog' event with no parameters.

When these events are emitted, they generate logs on the Ethereum blockchain. The logs can be observed and searched for using tools like Ethereum explorers (e.g., Etherscan). The 'indexed' parameter in the 'Log' event allows for more efficient filtering when querying logs based on the specified parameter