Create CallingParentContract.sol
This code illustrates a scenario involving multiple inheritance and the use of the super keyword.

Contracts A, B, C:
   - Contract A defines two functions 'foo' and 'bar', emitting log messages when called.
   - Contracts B and C inherit from A, each overriding both 'foo' and 'bar' and emitting their log messages before calling the corresponding function from A using 'A.foo()' or 'super.bar()'.

Contract D:
   - Contract D inherits from both B and C.
   - Overrides both 'foo' and 'bar' functions, calling 'super.foo()' and 'super.bar()' to ensure that the functions from the correct parent contracts are called.
   - The 'foo' function in D will call 'C.foo' and then 'A.foo', while the bar function in D will call 'C.bar', B.bar, and finally 'A.bar'.

When you interact with contract D by calling 'foo' and 'bar', you can observe the transaction logs to trace the sequence of function calls through the inheritance hierarchy. This example demonstrates the order of function execution when dealing with multiple inheritance and the use of 'super' to explicitly specify the parent contract from which to invoke a function.