Create FunctionSelector.sol
This contract has a function called 'getSelector', which takes a function signature as a string and returns its corresponding function selector. Function selectors are the first four bytes of the Keccak-256 hash of the function signature.

Example:

Input:
"transfer(address,uint256)"

Execution:
The 'getSelector' function calculates the Keccak-256 hash of the function signature: keccak256(bytes("transfer(address,uint256)")).
It then takes the first four bytes of the hash, converting it to a bytes4 type.

Output:
The function selector for the given function signature.

Usage:

You can use this contract to obtain the function selector for different function signatures.

This can be useful, especially when interacting with other contracts or creating transactions with specific function calls.